// apps/web/src/api-client/model/adminDashboardPopularRoute.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */

export interface AdminDashboardPopularRoute {
    originCity: string;
    destinationCity: string;
    /**
     * @minimum 0
     * @maximum 9007199254740991
     */
    count: number;
}